PCL_final/backend/requirements.txt
# AquaHub Backend Requirements
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-dotenv==1.0.0
